refactor(combat): remove unused ParticipantMother imports and adjust package structure for consistency
