+++ domain/src/test/java/com/arcathoria/combat/CombatTest.java
@@ -2,7 +2,6 @@ package com.arcathoria.combat;
 
 import com.arcathoria.combat.vo.CombatId;
 import com.arcathoria.combat.vo.CombatTurn;
-import com.arcathoria.combat.vo.ParticipantMother;
 import org.junit.jupiter.api.Test;
 
 import java.util.UUID;
