+++ domain/src/test/java/com/arcathoria/combat/MeleeCombatActionStrategyTest.java
@@ -1,7 +1,6 @@
 package com.arcathoria.combat;
 
 import com.arcathoria.combat.vo.CombatTurn;
-import com.arcathoria.combat.vo.ParticipantMother;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
