+++ domain/src/test/java/com/arcathoria/combat/ParticipantTest.java
@@ -1,6 +1,5 @@
-package com.arcathoria.combat.vo;
+package com.arcathoria.combat;
 
-import com.arcathoria.combat.Participant;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
