+++ shared-test-helpers/src/main/java/com/arcathoria/combat/CombatSnapshotMother.java
@@ -2,7 +2,6 @@ package com.arcathoria.combat;
 
 import com.arcathoria.combat.vo.CombatId;
 import com.arcathoria.combat.vo.CombatTurn;
-import com.arcathoria.combat.vo.ParticipantMother;
 
 import java.util.UUID;
